quell warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9663 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/LinkManager.cxx
@@ -211,7 +211,7 @@ void LinkManager::findTelesByName(const std::string& name,
   }
 
   // make the trailing face specification case-independent
-  const unsigned int last = glob.size() - 1;
+  const unsigned int last = (const unsigned int)glob.size() - 1;
   if ((glob[last] == 'F') || (glob[last] == 'B')) {
     glob[last] = tolower(glob[last]);
   }
